requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
pandas
sentence-transformers
langchain-text-splitters
python-box
google-cloud-firestore
langchain-google-firestore
flask